[nobank] move dashboard activity class to dashboard package.
