+++ NobankCore/noBank/src/main/java/com/hunk/nobank/NConstants.java
@@ -4,6 +4,8 @@ import android.content.Intent;
 import android.view.View;
 import android.widget.AdapterView;
 
+import com.hunk.nobank.activity.dashboard.DashboardActivity;
+
 public class NConstants {
     public final static int ANIMATION_DURATION_MEDIUM = 2000;
 
@@ -33,7 +35,7 @@ public class NConstants {
      *  @see
      *  com.hunk.nobank.activity.RootActivity#onNewIntent(Intent)
      *  @see
-     *  com.hunk.nobank.activity.DashboardActivity#onBackPressed()
+     *  DashboardActivity#onBackPressed()
      */
     public static final String INTENT_EXTRA_IS_EXIT = "INTENT_EXTRA_IS_EXIT";
 }
